+++ pkg/instrumented/instrumentedHandler.go
@@ -1,9 +1,25 @@
 package instrumented
 
 import (
+	"errors"
 	"net/http"
 
 	"github.com/gesundheitscloud/go-monitoring/prom"
+	"github.com/gesundheitscloud/go-svc/pkg/logging"
+)
+
+var (
+	//LatencyBuckets define buckets for histogram of HTTP request/reply latency metric - in seconds
+	LatencyBuckets = []float64{.0001, .001, .01, .1, .25, .5, 1, 2.5, 5, 10}
+	//SizeBuckets define buckets for histogram of HTTP request/reply size metric - in bytes
+	SizeBuckets = []float64{16, 32, 64, 128, 256, 512, 1024, 5120, 20480, 102400, 512000, 1000000, 10000000}
+	//DefaultInstrumentOptions hold options (API-path-specific) for HTTP instrumenter - record request size and response size
+	DefaultInstrumentOptions = []prom.Option{prom.WithReqSize, prom.WithRespSize}
+	//DefaultInstrumentInitOptions hold initialization options (API-handler-specific) for HTTP instrumenter - definitions of histogram buckets
+	DefaultInstrumentInitOptions = []prom.InitOption{
+		prom.WithLatencyBuckets(LatencyBuckets),
+		prom.WithSizeBuckets(SizeBuckets),
+	}
 )
 
 // HandlerFactory stores default settings and produces  handlers
@@ -23,8 +39,22 @@ func NewHandlerFactory(subsystemName string, defaultInitOpts []prom.InitOption, 
 }
 
 // NewHandler produces an Handler with default values set
-func (ihf *HandlerFactory) NewHandler(handlerName string, extraOpts ...prom.Option) *Handler {
-	return newHandler(ihf.subsystemName, handlerName, ihf.defaultInitOptions, ihf.defaultOptions)
+func (ihf *HandlerFactory) NewHandler(handlerName string, extraOpts ...interface{}) *Handler {
+	initOptions := ihf.defaultInitOptions
+	options := ihf.defaultOptions
+
+	for _, o := range extraOpts {
+		switch opt := o.(type) {
+		case prom.InitOption:
+			initOptions = append(initOptions, opt)
+		case prom.Option:
+			options = append(options, opt)
+		default:
+			logging.LogWarningf(errors.New("unknown instrumentation option type"), "change option to a known option")
+		}
+
+	}
+	return newHandler(ihf.subsystemName, handlerName, initOptions, options)
 }
 
 // Handler holds instrumentation object for HTTP handlers