+++ pkg/middlewares/auth_test.go
@@ -1,0 +1,105 @@
+package middlewares
+
+import (
+	"crypto/rand"
+	"crypto/rsa"
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/gesundheitscloud/go-svc/pkg/instrumented"
+
+	"github.com/dgrijalva/jwt-go"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestServiceSecret(t *testing.T) {
+	validAuthHeader := "service-secret"
+
+	handlerFactory := instrumented.NewHandlerFactory("d4l", instrumented.DefaultInstrumentInitOptions, instrumented.DefaultInstrumentOptions)
+	auth := NewAuth(validAuthHeader, nil, handlerFactory)
+	tests := []struct {
+		name              string
+		AuthHeaderContent string
+		expectedStatus    int
+	}{
+		{"valid Service Auth", validAuthHeader, http.StatusOK},
+		{"invalid Service Auth", "random", http.StatusUnauthorized},
+	}
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			req, _ := http.NewRequest(http.MethodGet, "/route", nil)
+			req.Header.Add(AuthHeaderName, tt.AuthHeaderContent)
+			res := httptest.NewRecorder()
+
+			// target handler after auth check
+			handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
+			authMiddleware := auth.ServiceSecret(handler)
+			authMiddleware.ServeHTTP(res, req)
+			assert.Equal(t, tt.expectedStatus, res.Code)
+		})
+	}
+}
+
+func TestJwt(t *testing.T) {
+	// create rsa keypait for testing
+	reader := rand.Reader
+	bitSize := 512
+	privateKey, _ := rsa.GenerateKey(reader, bitSize)
+
+	handlerFactory := instrumented.NewHandlerFactory("d4l", instrumented.DefaultInstrumentInitOptions, instrumented.DefaultInstrumentOptions)
+	auth := NewAuth("", &privateKey.PublicKey, handlerFactory, AuthWithLatencyBuckets([]float64{4, 8, 16}), AuthWithSizeBuckets([]float64{4, 8, 16}))
+	claims := &jwt.StandardClaims{
+		Issuer: "test",
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodRS256, claims)
+	signedToken, _ := token.SignedString(privateKey)
+
+	tests := []struct {
+		name              string
+		AuthHeaderContent string
+		expectedStatus    int
+	}{
+		{"valid Service Auth", signedToken, http.StatusOK},
+		{"invalid Service Auth", "random", http.StatusUnauthorized},
+	}
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			req, _ := http.NewRequest(http.MethodGet, "/route", nil)
+			req.Header.Add(AuthHeaderName, fmt.Sprintf("Bearer %s", tt.AuthHeaderContent))
+			res := httptest.NewRecorder()
+
+			handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
+			authMiddleware := auth.JWT(handler)
+			authMiddleware.ServeHTTP(res, req)
+			assert.Equal(t, tt.expectedStatus, res.Code)
+		})
+	}
+}
+
+func TestGetAuthSecret(t *testing.T) {
+	tests := []struct {
+		name              string
+		authHeaderContent string
+		expectedSecret    string
+	}{
+		{"Service secret without prefix", "secret", "secret"},
+		{"Service secret with prefix", "Bearer anothersecret", "anothersecret"},
+	}
+
+	auth := &Auth{}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			req, _ := http.NewRequest(http.MethodGet, "/route", nil)
+			req.Header.Add(AuthHeaderName, tt.authHeaderContent)
+
+			authToken, _ := auth.getAuthSecret(req)
+			assert.Equal(t, authToken, tt.expectedSecret)
+		})
+	}
+}